backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snowbird-backend"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy==2.0.30",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==3.2.2",
    "python-jose[cryptography]==3.3.0",
    "cryptography==42.0.7",
    "alpaca-py==0.26.0",
    "apscheduler==3.10.4",
    "redis==5.0.4",
    "python-multipart==0.0.9",
    "httpx==0.27.0",
    "python-dateutil==2.9.0",
    "pytz==2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "freezegun==1.5.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
